Rename methods, move to seperate files and use import instead of require
