+++ src/transforms/literalToObject.js
@@ -1,0 +1,17 @@
+import * as tt from '@babel/types';
+
+/**
+ * Converts a member expression from bracket notation to object notation and
+ * returns a new MemberExpression with the changes.
+ *
+ * @param {Node} leftNode - The left expression in the node
+ * @return {MemberExpression} MemberExpression
+ *
+ * @example Class.prototype['methodName'] -> Class.prototype.methodName
+ */
+export default function literalToObject(leftNode) {
+  const { property, object } = leftNode;
+  const newProperty = tt.identifier(property.value);
+  const newAssignment = tt.memberExpression(object, newProperty, false);
+  return newAssignment;
+}
