+++ src/transforms/utils/isProtoLiteralAssignment.js
@@ -1,0 +1,20 @@
+import * as tt from '@babel/types';
+
+/**
+ * Checks if there is an assignment expression for a class prototype using
+ * bracket notation.
+ *
+ * @param {Node} node The AST node from babels parser
+ *
+ * @example Class.prototype['methodName']
+ */
+export default function isProtoLiteralAssignment(node) {
+  if (tt.isAssignmentExpression(node) && tt.isLiteral(node.left.property)) {
+    const { property, object } = node.left;
+    if (object.property) {
+      return object.property.name === "prototype";
+    }
+    return false;
+  }
+  return false;
+}
