+++ src/transforms/utils/isProtoLiteralVar.js
@@ -1,0 +1,22 @@
+import * as tt from '@babel/types';
+/**
+ * Checks if there is a variable declaration for a class prototype using bracket
+ * notation.
+ *
+ * @param {Node} node The AST node from babels parser
+ *
+ * @example var alias = Class.prototype['methodName']
+ */
+export default function isProtoLiteralVar(node) {
+  if (tt.isVariableDeclaration(node)) {
+    const declaration = node.declarations[0];
+    if (
+      tt.isMemberExpression(declaration.init) &&
+      tt.isLiteral(declaration.init.property)
+    ) {
+      const { property } = declaration.init.object;
+      return property && property.name === "prototype";
+    }
+  }
+  return false;
+}
